Clean disk cache when app enter background or will terminate;

+++ Example/CZCacheExample/CZCacheExample/CZCache/DiskCacheSupport/CZKVStore.h
@@ -46,8 +46,8 @@ NS_ASSUME_NONNULL_BEGIN
 - (BOOL)removeItemsWithCountLimit:(NSInteger)count;
 - (BOOL)removeItemsEarlierThanDate:(NSInteger)date;
 
-- (int)totalItemsSize;
-- (int)totalItemsCount;
+- (NSInteger)totalItemsSize;
+- (NSInteger)totalItemsCount;
 
 @end
